Merge branch 'js/am-3-merge-recursive-direct' into jch
The merge_recursive_generic() function has been made a bit safer to
call from inside a process.  "git am -3" was taught to make a direct
call to the function when falling back to three-way merge.

Being able to make a direct call would be good in general, but as a
performance thing, the change needs to be backed up by numbers.

I haven't gone through the "gently" change with fine toothed comb;
I can see that the change avoids calling die(), but I haven't made
sure that the program states (e.g. what's in the in-core index) are
adjusted sensibly when it returns to the caller instead of dying,
or the codepaths that used to die() are free of resource leaks.
The original code certainly did not care the program states at the
point of dying exactly because it knew it is going to exit, but now
they have to care, and they need to be audited.

* js/am-3-merge-recursive-direct:
  am: make a direct call to merge_recursive
  merge_recursive_options: introduce the "gently" flag

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>