sequencer: remember special error conditions
A cherry-pick can fail in surprising ways. For example, if it would
overwrite an untracked file, it does not even start to do anything, let
alone record that it is cherry-picking.

To discern such cases from regular error conditions where the user
simply has to resolve a merge conflict and then can call `git commit` to
finalize the cherry-pick (reusing the original commit message), such a
"really fatal" error is indicated by using an exit code != 1.

This means that we have to remember such an exit code by OR'ing later
error conditions.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>