rebase -i: cross-validate rebase--helper's functionality
Inspired by GitHub's "Scientist" library (and by this developer's
validation of the builtin replacement for RCS merge, back in the days),
we run things *twice*: once with the original shell script code, and
once with the new rebase--helper method.

Note that this will take a toll on users for now: merge conflicts and
rewords have to be addressed twice, too, and care needs to be taken that
exec commands (and copying commit notes) can be performed twice; If not,
the environment variable GIT_USE_REBASE_HELPER=true needs to be set (or,
in case of breakage, it needs to be set to "false" before rerunning the
rebase).

Please note that for the same reason, some tests fail with merge
conflicts, or specifically test amending commits via "reword" or "edit"
commands. Naturally, these tests cannot be cross-validated easily,
because the same "interactive" commands would have to be run again. So
let's just run the rebase--helper method to verify that these tests
still pass.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>