WIP: perf: add a script to extrapolate perf numbers
When working toward improving performance, ideally the tests are *not*
performed on a large-scale repository: it slows down the turn-around, and
hence also the development time.

Instead, we try to extrapolate from several smaller-scale scenarios what
the performance characteristics of the large-scale repository would be.

This patch introduces a Perl script that extrapolates such performance
numbers by fitting a function with constant, linear, quadratic and
n*log(n) components.

Note: in a preliminary experiment, this method was highly susceptible to
normal fluctuations and fitted *negative* factors (which of course makes
no sense). A possible future course of action would be to replace the
regular least-squares fit by an iterative non-negative least-squares fit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>